implementacao da listagem de comics por usuario. Consumindo os comics da api da marvel e armazenando no bano local e iniciao das validações de comic

+++ backend/src/main/java/br/com/zup/desafio/marvel/api/MarvelComicsClient.java
@@ -13,7 +13,7 @@ public interface MarvelComicsClient {
 	
 	@GetMapping("/{id}")
 	ResponseEntity<ComicsResponse> findComicsPorId(
-			@PathVariable Integer id,
+			@PathVariable Long id,
 			@RequestParam String ts,
             @RequestParam String apikey,
             @RequestParam String hash);
